+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * 5.0.0.SNAPSHOT.20220000: Complete review of the plugin. Updated focussed mainly on server side/microservices architecture.
-
+    * **UTILS**:
+        * **IMPROVEMENT** :-1: : Cleaned `utils`.